+++ src/impls/mod.rs
@@ -8,6 +8,9 @@
 pub mod hyper_log_log;
 pub use hyper_log_log::{HyperLogLog, HyperLogLogBuilder, HyperLogLogError, HyperLogLogHelper};
 
+pub mod hyper_log_log8;
+pub use hyper_log_log8::{HyperLogLog8, HyperLogLog8Builder};
+
 mod slice_estimator_array;
 pub use slice_estimator_array::*;